requirements.txt
# --- Core ---
pandas>=2.2.0
numpy>=1.26.0

# --- Deep Learning ---
torch>=2.2.0
transformers>=4.44.0
accelerate>=0.33.0
datasets>=2.0.0

# --- Metrics & Evaluation ---
seqeval>=1.2.2
scikit-learn>=1.5.0

# --- Utilities ---
tqdm>=4.66.0
rich>=13.7.0
pyyaml>=6.0.1

# --- Data Handling ---
matplotlib>=3.8.0
seaborn>=0.13.0
dvc

# --- Optional but Recommended ---
notebook>=7.0.0         # Jupyter support